requirements.txt
openai
langchain
pinecone-client
pyautocad
numpy
pandas
kittycad==0.7.4
python-dotenv
pygltflibpyvista
ollama==0.3.3 # Local inference
chromadb==0.5.20 # Vector Database
sentence-transformers==3.3.1 # CrossEncoder Re-ranking
streamlit==1.40.1 # Application UI
PyMuPDF==1.24.14 # PDF Document loader
langchain-community==0.3.7 # Utils for text splitting
torch
pyyaml
PyPDF2
beautifulsoup4
lxml
